requirements.txt
anyio==3.6.2
APScheduler==3.10.1
backports.zoneinfo==0.2.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.94.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
mysql-connector-python==8.0.32
passlib==1.7.4
protobuf==3.20.3
pyasn1==0.4.8
pydantic==1.10.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
starlette==0.26.0.post1
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
